:sparkles: Update multiple dependencies and add Logger interface
This commit updates multiple dependencies in the go.mod and go.sum files.
It also introduces a new Logger interface in logger.go for structured logging.
The option to use a custom logger is now added to the NewOtq function in otq.go. Unnecessary components in .idea/misc.xml have been removed, and a gitignore rule has been added in .idea/.gitignore.